UC-5 Success flow Test Added
Added the successful flow test case for UC-5, Modify Room Properties.
Assertions on the modified values are made, and the decrease in the
total number of enabled checkmarks is asserted to be 1 (the test case
sets the room to disabled). The rest assumes that the room added for
modification is located in the second row.