UC-5 Success Flow Test Added
Added the successful flow system test of UC-5, Modify Room Properties.
Assertions are made to ensure that the values of the room we
successfully changed. The test assumes that the room created of
modification for the purposes of the test is added to the second row of
the rooms table